Create README.md
It is possible to split an ERC-1155 token, which is a type of NFT, into multiple smaller ERC-20 tokens using a smart contract. The smart contract would be responsible for dividing the ERC-1155 token into the desired number of smaller ERC-20 tokens and allocating them to the appropriate parties. Each ERC-20 token would represent a partial ownership of the original ERC-1155 token.